* Fixed subtle bug in config file management involving references to 'guint8' and 'guint16' types being casted to int references. * Tagged as rc2
Please make sure you know what you're doing (read the big warning header
I just added) when adding config options!

Waiting for a couple of DC changes for rc3.

-- Christoph

+++ src/include/sci_conf.h
@@ -94,8 +94,8 @@ typedef struct {
 	
 	pcmout_driver_t *pcmout_driver; /* the pcm driver to use */
 	sound_server_t *sound_server; /* The sound server */
-        guint16 pcmout_rate;  /* Sample rate */
-        guint8 pcmout_stereo;  /* Stereo? */
+        int pcmout_rate;  /* Sample rate */
+        int pcmout_stereo;  /* Stereo? */
 
 	char *module_path; /* path to directories modules are loaded from */
